[Update] Update auth.js to save user_id from db
Previously auth.js was saving the user’s Facebook ID to chrome.storage.sync. This is because the rest of the app uses that ID to identify API requests. However, we’ve decided to use the user’s `user_id` in our database as the identifier. This commit is the beginning of that transition.