[Update] Store entire user obj in chrome.storage
Previously we were only storing the user’s id in our database in chrome.storage. Now, in addition to the user’s id in our database, we store their full name, email, and picture URL. The description property of the user object is currently present but commented out, in anticipation of changes to the backend.

Also removed some console.logs in `test.js` and `main.js`.