[Refactor] Search GET->child component,gulp-notify
Moved AJAX call for searching for people out of the `handleSubmit` function in `FeedSearchView` and into the `componentWillReceiveProps` method of the `FeedSearchList` component.

Also added `gulp-notify` to `gulpfile.js` so that errors don’t just silently appear in the Terminal and cause gulp to crash. The new functionality is that any errors are reported via the OS notification system and that gulp continues to watch for any changes in your files.